Add files via upload
The RollingPaper is your gateway to "The High Club Token" ($THC), where we unite a laid-back approach with groundbreaking crypto innovation. Inspired by a developer who's "All the Time High" thanks to a certain green companion, marijuana, THC Token offers a unique and enjoyable crypto experience. Key features include a solid contract, locked liquidity, and smart taxation. The tokenomics comprise a total supply of 10 million tokens, a burning pipe, and a roadmap with three key phases. In the world of THC Token, innovation, transparency, and constant inspiration reign. Join us for a new high in the digital space! 🚀🌿